[CSL-1346] expect response after inv/data
A ResMsg is expected following the DataMsg if the peer send a ReqMsg. It
includes a Bool to indicate whether the data was successfully processed,
in particular whether a transaction was successfully included in the
peer's mempool.

I judged the design in here to be the shortest path to a working
solution but I don't believe it's ideal. We may want to consider opening
up time-warp-nt to allow for conversations on heterogeneous send/recv
types so that we don't have to create these unions InvOrData, ReqOrRes
and explicitly handle the possibility that the peer sends a value of
the wrong type.